perf: add auto check if we can remove filter (#6283)

+++ src/service/search/cluster/flight.rs
@@ -979,7 +979,7 @@ pub async fn get_inverted_index_file_list(
     req.stream_type = StreamType::Index;
     query.sql = sql;
     query.from = 0;
-    query.size = QUERY_WITH_NO_LIMIT;
+    query.size = QUERY_WITH_NO_LIMIT as i32;
     query.track_total_hits = false;
     query.uses_zo_fn = false;
     query.query_fn = "".to_string();
